開始時の魔法陣の演出、カメラの移動

+++ WallMaker/GameSources/UISprite.cpp
@@ -176,7 +176,7 @@ namespace basecross{
 	void WallStock::OnUpdate()
 	{
 		auto player_share = GetStage()->GetSharedGameObject<Player>(WstringKey::ShareObj_Player);
-		float playerWallStock = player_share->GetWallStock();
+		int playerWallStock = player_share->GetWallStock();
 		if (playerWallStock >= m_WallNum)
 		{
 			////F‚ğ–ß‚·
